lawbroke-backend-main/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==4.2.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
attrs==25.1.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
databases==0.9.0
datasets==3.1.0
dill==0.3.8
entrypoints==0.4
ethnicolr==0.18.4
fastapi==0.115.5
filelock==3.16.1
Flask==3.0.3
Flask-RESTful==0.3.10
frozenlist==1.5.0
fsspec==2024.9.0
gender-guesser==0.4.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.13.5
networkx==3.4.2
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
passlib==1.7.4
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyarrow==18.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
pydeps==3.0.0
Pygments==2.18.0
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.28.1
setuptools==75.5.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
starlette==0.41.3
stdlib-list==0.11.0
streamlit==1.40.1
streamlit-aggrid==1.0.5
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
toolz==1.0.0
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
uvicorn==0.32.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.17.2
